Extract RootLayout and use it in debugger page
