+++ apps/store/src/app/[locale]/DebugError.tsx
@@ -4,7 +4,7 @@ import { useSearchParams } from 'next/navigation'
 export const DebugError = () => {
   const searchParams = useSearchParams()
   if (process.env.VERCEL_ENV !== 'production' && searchParams?.get('rootError')) {
-    throw new Error('root layout error')
+    throw new Error('Debug: root layout error')
   }
   return null
 }
