+++ apps/store/src/app/[locale]/layout.tsx
@@ -2,13 +2,9 @@
 // next.config.js - { compiler: { emotion: true } }
 /** @jsxImportSource react */
 
-import { Provider as JotaiProvider } from 'jotai'
-import { ReactNode, Suspense } from 'react'
-import { NextAppDirEmotionCacheProvider } from 'tss-react/next/appDir'
-import globalCss from 'ui/src/global.css'
+import { ReactNode } from 'react'
 import { theme } from 'ui'
 import { fetchGlobalProductMetadata } from '@/components/LayoutWithMenu/fetchProductMetadata'
-import { ApolloWrapper } from '@/services/apollo/app-router/ApolloWrapper'
 import { getClient } from '@/services/apollo/app-router/rscClient'
 import { ShopSessionProvider } from '@/services/shopSession/ShopSessionContext'
 import { GlobalStory } from '@/services/storyblok/storyblok'
@@ -16,22 +12,14 @@ import {
   fetchStoryblokCacheVersion,
   getStoryBySlug,
 } from '@/services/storyblok/storyblok.serverOnly'
-import { contentFontClassName } from '@/utils/fonts'
 import { locales } from '@/utils/l10n/locales'
-import { getLocaleOrFallback } from '@/utils/l10n/localeUtils'
 import { RoutingLocale } from '@/utils/l10n/types'
 import { ORIGIN_URL } from '@/utils/PageLink'
-import { DebugError } from './DebugError'
 import { initTranslationsServerSide } from './i18n'
 import { ProductMetadataProvider } from './ProductMetadataProvider'
+import { RootLayout } from './RootLayout'
 import { StoryblokLayout } from './StoryblokLayout'
-import StoryblokProvider from './StoryblokProvider'
 import { TranslationsProvider } from './TranslationsProvider'
-
-// Trick compiler into thinking we need global.css import for anything other than side effects
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const noop = (val: any) => {}
-noop(globalCss)
 
 export type LocalizedLayoutProps<P = unknown> = P & {
   children: ReactNode
@@ -47,37 +35,23 @@ const Layout = async ({ children, params: { locale } }: LocalizedLayoutProps) =>
   ])
 
   return (
-    <html lang={getLocaleOrFallback(locale).htmlLang}>
-      <head>
-        <meta name="theme-color" content={theme.colors.light} />
-      </head>
-      <body className={contentFontClassName}>
-        <NextAppDirEmotionCacheProvider options={{ key: 'css' }}>
-          <Suspense>
-            <DebugError />
-          </Suspense>
-          <ApolloWrapper>
-            <StoryblokProvider>
-              <JotaiProvider>
-                <TranslationsProvider locale={locale} resources={resources}>
-                  <ProductMetadataProvider locale={locale} productMetadata={productMetadata}>
-                    <ShopSessionProvider>
-                      <StoryblokLayout globalStory={globalStory}>{children}</StoryblokLayout>
-                    </ShopSessionProvider>
-                  </ProductMetadataProvider>
-                </TranslationsProvider>
-              </JotaiProvider>
-            </StoryblokProvider>
-          </ApolloWrapper>
-        </NextAppDirEmotionCacheProvider>
-      </body>
-    </html>
+    <RootLayout locale={locale}>
+      <TranslationsProvider locale={locale} resources={resources}>
+        <ProductMetadataProvider locale={locale} productMetadata={productMetadata}>
+          <ShopSessionProvider>
+            <StoryblokLayout globalStory={globalStory}>{children}</StoryblokLayout>
+          </ShopSessionProvider>
+        </ProductMetadataProvider>
+      </TranslationsProvider>
+    </RootLayout>
   )
 }
 
 export const generateStaticParams = () => {
   return Object.values(locales).map(({ routingLocale }) => ({ locale: routingLocale }))
 }
+
+export const dynamicParams = false
 
 export default Layout
