+++ apps/store/src/app/[locale]/test/ClientComponent.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useAtomValue } from 'jotai'
-import { useSearchParams } from 'next/navigation'
+import { notFound, useSearchParams } from 'next/navigation'
 import { useTranslation } from 'next-i18next'
 import { productsMetadataAtom } from '@/components/LayoutWithMenu/productMetadataAtom'
 import { useAppRouterLocale } from '../useAppRouterLocale'
@@ -10,7 +10,9 @@ export const ClientComponent = () => {
   const { t } = useTranslation('purchase-form')
   const searchParams = useSearchParams()
   if (searchParams?.get('error')) {
-    throw new Error('test error')
+    throw new Error('Debug: client component error')
+  } else if (searchParams?.get('notFound')) {
+    throw notFound()
   }
   const locale = useAppRouterLocale()
   // Read jotai value from provider
