--- apps/store/src/app/[locale]/test404/page.ts
@@ -1,6 +1,0 @@
-import { notFound } from 'next/navigation'
-
-// Provide a way to trigger 404 in app router
-export default function Test404Page() {
-  throw notFound()
-}
