+++ apps/store/src/app/debugger/page.tsx
@@ -1,28 +1,31 @@
-'use client'
-
-import styled from '@emotion/styled'
 import Head from 'next/head'
+import { RootLayout } from '../[locale]/RootLayout'
 import { CreateSessionForm } from './CreateSessionForm'
 
 const Page = () => {
   return (
-    <Wrapper>
-      <Head>
-        <meta name="robots" content="noindex,nofollow" />
-      </Head>
-
-      <CreateSessionForm />
-    </Wrapper>
+    <RootLayout locale={'se-en'}>
+      <div style={wrapperStyle}>
+        <Head>
+          <meta name="robots" content="noindex,nofollow" />
+        </Head>
+        <CreateSessionForm />
+      </div>
+    </RootLayout>
   )
 }
 
 export default Page
 
-const Wrapper = styled.div({
+export const metadata = {
+  robots: 'noindex, nofollow',
+}
+
+const wrapperStyle = {
   display: 'flex',
   justifyContent: 'center',
   alignItems: 'center',
   height: '100%',
   maxWidth: '20rem',
   marginInline: 'auto',
-})
+}
